# docs/canonical/ATMAN_CANONICAL_DOCS_v0_2_1/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Atman Platform API
  version: 0.2.1
  description: Canonical API surface for ATMAN_REPO_v0_3.
servers:
  - url: http://localhost:8000
paths:
...
  /sources:
    post:
      summary: Register a source
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceCreateRequest'
      responses:
        '200':
          description: Source registered
    get:
      summary: List sources
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Source list
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SourceCreateRequest:
      type: object
      required: [title, source_type, language]
      properties:
        title: { type: string }
        source_type: { type: string, enum: [pdf, epub, txt, url, web, manual] }
        language: { type: string, enum: [hi, sa, en, mixed] }
        rights_status: { type: string, default: unknown }
...